Data Visualization of the Movies in the Data
This a quick Tableau visual of the movies in the data with at least 50 views. 

On the X-Axis is the Average Rating amongst every person who saw that particular movie, and the Y Axis has the amount of times that movie was watched. 

You can see we have them blocked off into four quadrants
- Box Office (High Average Rating, High Times Watched)
- Sleepers (High Average Rating, Low Times Watched)
- Flops (Low Average Rating, Low Times Watched)
- Over Hyped (Low Average Rating, High Times Watched)